Move components to shared-components
