+++ packages/wallet/src/redux/protocols/dispute/challenger/components/defund-or-not.tsx
@@ -13,12 +13,8 @@ export default class DefundOrNot extends React.PureComponent<Props> {
     return (
       <Fragment>
         <h1>Challenge timed out!</h1>
-
-        <p>
-          The challenge timed out. Channel {channelId} is now finalized -- would you like to defund
-          it?
-        </p>
-
+        The challenge timed out. Channel <div className="channel-address">{channelId}</div> is now
+        finalized -- would you like to defund it?
         <Button onClick={deny}>No</Button>
         <Button onClick={approve}>Defund</Button>
       </Fragment>
