+++ packages/wallet/src/redux/protocols/funding/player-b/container.tsx
@@ -8,8 +8,8 @@ import * as states from './states';
 import { unreachable } from '../../../../utils/reducer-utils';
 import { TwoPartyPlayerIndex } from '../../../types';
 import ApproveStrategy from '../../../../components/funding/approve-strategy';
-import WaitForOtherPlayer from '../../../../components/wait-for-other-player';
-import AcknowledgeX from '../../../../components/acknowledge-x';
+import WaitForOtherPlayer from '../../shared-components/wait-for-other-player';
+import AcknowledgeX from '../../shared-components/acknowledge-x';
 import { FundingStrategy } from '..';
 import { ActionDispatcher } from '../../../utils';
 import { IndirectFunding } from '../../indirect-funding/container';
