+++ packages/wallet/src/redux/protocols/shared-components/acknowledge-x.tsx
