+++ packages/wallet/src/redux/protocols/shared-components/approve-x.tsx
