+++ packages/wallet/src/redux/protocols/shared-components/status-bar-layout.tsx
