+++ packages/wallet/src/redux/protocols/shared-components/transaction-failed.tsx
