+++ packages/wallet/src/redux/protocols/shared-components/wait-for-other-player.tsx
