+++ packages/wallet/src/redux/protocols/shared-components/yes-or-no.tsx
